--- database/connector.py
@@ -1,19 +1,0 @@
-from sqlalchemy import create_engine
-from sqlalchemy.orm import scoped_session, sessionmaker
-from sqlalchemy.ext.declarative import declarative_base
-
-from config import DATABASE_CONFIG
-
-engine = create_engine(DATABASE_CONFIG["connection_string"], convert_unicode=True)
-db_session = scoped_session(sessionmaker(autocommit=DATABASE_CONFIG["autocommit"],
-                                         autoflush=DATABASE_CONFIG["autoflush"],
-                                         bind=engine))
-Base = declarative_base()
-Base.query = db_session.query_property()
-
-
-def init_db():
-    # import all modules here that might define models so that
-    # they will be registered properly on the metadata.  Otherwise
-    # you will have to import them first before calling init_db()
-    Base.metadata.create_all(bind=engine)
